feat(ai-evaluator): remove toggle button approach
