+++ libs/data-access/api/src/lib/trpc/routers/text-evaluation.router.ts
@@ -2,9 +2,6 @@ import { TRPCError } from "@trpc/server";
 import { authProcedure, t } from "../trpc";
 import { database } from "@self-learning/database";
 import { z } from "zod";
-
-// ─── Types ────────────────────────────────────────────────────────────────────
-
 interface LlmConfig {
 	serverUrl: string;
 	apiKey: string | null;
@@ -16,10 +13,6 @@ type Message = {
 	content: string;
 };
 
-// ─── Response schema from the LLM server ─────────────────────────────────────
-// Matches the same shape used in ai-tutor.router.ts so both routers stay
-// compatible with the same LLM server.
-
 const llmApiResponseSchema = z.object({
 	message: z.object({
 		content: z.string(),
@@ -27,27 +20,17 @@ const llmApiResponseSchema = z.object({
 	})
 });
 
-// ─── Input / Output schemas ───────────────────────────────────────────────────
-
 const evaluateInputSchema = z.object({
-	/** The question text shown to the student */
 	questionStatement: z.string().min(1),
-	/** The teacher's sample solution or concept list (as plain text) */
 	solutionOrConcepts: z.string().min(1),
-	/** 0–100: minimum coverage required to pass */
 	passingThreshold: z.number().int().min(0).max(100),
-	/** The raw text the student typed as their answer */
 	studentAnswer: z.string().min(1)
 });
 
 const evaluateOutputSchema = z.object({
-	/** One of the four verdict levels */
 	verdict: z.enum(["correct", "partially-correct", "partially-wrong", "wrong"]),
-	/** Short explanation in the same language as the question */
 	feedback: z.string()
 });
-
-// ─── Router ──────────────────────────────────────────────────────────────────
 
 export const textEvaluationRouter = t.router({
 	/**
@@ -57,25 +40,25 @@ export const textEvaluationRouter = t.router({
 	 *  1. Fetching the LLM config from the database
 	 *  2. Building the system prompt (security-sensitive: must never leak the solution)
 	 *  3. Building the user message with all evaluation context
-	 *  4. Calling the LLM with temperature 0 (FR-09)
-	 *  5. Parsing and validating the structured JSON response (FR-11)
+	 *  4. Calling the LLM with temperature 0
+	 *  5. Parsing and validating the structured JSON response
 	 *  6. Returning a typed { verdict, feedback } object
 	 *
 	 * All LLM communication errors are re-thrown as TRPCErrors so the client
-	 * (component.tsx) can catch them uniformly and show the fallback message (FR-16).
+	 * (component.tsx) can catch them uniformly and show the fallback message.
 	 */
 	evaluate: authProcedure
 		.input(evaluateInputSchema)
 		.output(evaluateOutputSchema)
 		.mutation(async ({ input, ctx }) => {
 			// Step 1: Fetch LLM configuration from DB
-			// Throws NOT_FOUND if no config exists — caught by component.tsx (FR-12)
+			// Throws NOT_FOUND if no config exists — caught by component.tsx
 			const llmConfig = await fetchLlmConfig();
 
 			// Step 2: Build system prompt — evaluation-specific, never reveals solution
 			const systemPrompt = buildSystemPrompt();
 
-			// Step 3: Build the user message with all four required pieces of context (FR-07)
+			// Step 3: Build the user message with all four required pieces of context
 			const userMessage = buildUserMessage(input);
 
 			// Step 4: Send to LLM and get raw response string
@@ -86,7 +69,7 @@ export const textEvaluationRouter = t.router({
 
 			const rawContent = await sendEvaluationRequest(messages, llmConfig);
 
-			// Step 5: Parse and validate the structured JSON the LLM returned (FR-11)
+			// Step 5: Parse and validate the structured JSON the LLM returned
 			const parsed = parseLlmResponse(rawContent);
 
 			console.info("[TextEvaluationRouter] Evaluation complete", {
@@ -114,19 +97,17 @@ export const textEvaluationRouter = t.router({
 	})
 });
 
-// ─── System Prompt ────────────────────────────────────────────────────────────
-
 /**
  * Builds the system prompt for the LLM evaluator.
  *
- * Security rules encoded here (FR-10):
+ * Security rules encoded here:
  *  - Must NEVER reveal, hint at, paraphrase, or quote the sample solution
  *  - Feedback must only address what the student actually wrote
  *
- * Evaluation rules (FR-08, FR-11):
+ * Evaluation rules:
  *  - LLM decides pass/fail itself based on the threshold — we do not decide for it
  *  - Response must be pure JSON with exactly the fields { verdict, feedback }
- *  - Feedback must be in the same language as the question (FR-11)
+ *  - Feedback must be in the same language as the question
  */
 function buildSystemPrompt(): string {
 	return `You are a strict but fair academic evaluator for an e-learning platform.
@@ -144,7 +125,7 @@ You will receive:
 3. Choose the appropriate verdict based on the coverage and correctness of the student's answer:
    - "correct": The student fully understood and addressed the question. Coverage meets or exceeds the passing threshold.
    - "partially-correct": Meaningful understanding shown, but some important aspects are missing. Coverage is close to but just below the threshold.
-   - "partially-wrong": Some relevant knowledge shown, but significant parts are missing or incorrect.
+   - "partially-wrong": Some relevant knowledge shown, but significant parts incorrect.
    - "wrong": The answer does not address the question, is off-topic, or is too vague to evaluate.
 4. Write a short feedback (1–3 sentences) that explains your verdict.
 
@@ -160,11 +141,9 @@ You will receive:
 {"verdict": "<correct|partially-correct|partially-wrong|wrong>", "feedback": "<your feedback here>"}`;
 }
 
-// ─── User Message ─────────────────────────────────────────────────────────────
-
 /**
  * Builds the user-role message containing all four evaluation context pieces.
- * FR-07: question text, evaluation reference, threshold, and student answer must all be present.
+ * question text, evaluation reference, threshold, and student answer must all be present.
  */
 function buildUserMessage(input: z.infer<typeof evaluateInputSchema>): string {
 	return `## Question
@@ -180,19 +159,17 @@ ${input.passingThreshold}%
 ${input.studentAnswer}`;
 }
 
-// ─── LLM Communication ───────────────────────────────────────────────────────
-
 /**
  * Sends the evaluation request to the LLM server.
  *
  * Key differences from the AI Tutor's sendChatRequest:
- *  - temperature: 0  (FR-09 — deterministic, reproducible results)
+ *  - temperature: 0  (deterministic, reproducible results)
  *  - maxTokens: 300  (feedback is short: 1–3 sentences)
  *  - No streaming
  *
  * Throws TRPCError on any failure (timeout, HTTP error, invalid response shape).
  * The caller (evaluate procedure) propagates these to the client, which shows
- * the unified fallback message (FR-16).
+ * the unified fallback message.
  */
 async function sendEvaluationRequest(messages: Message[], config: LlmConfig): Promise<string> {
 	const TIMEOUT_MS = 30_000;
@@ -211,7 +188,7 @@ async function sendEvaluationRequest(messages: Message[], config: LlmConfig): Pr
 				messages,
 				model: config.defaultModel,
 				stream: false,
-				temperature: 0, // FR-09: must be 0 for evaluation
+				temperature: 0, // must be 0 for evaluation
 				maxTokens: 300 // short feedback only
 			}),
 			signal: controller.signal
@@ -260,8 +237,6 @@ async function sendEvaluationRequest(messages: Message[], config: LlmConfig): Pr
 	}
 }
 
-// ─── Response Parsing ─────────────────────────────────────────────────────────
-
 /**
  * Parses the raw string returned by the LLM into a typed { verdict, feedback } object.
  *
@@ -271,7 +246,7 @@ async function sendEvaluationRequest(messages: Message[], config: LlmConfig): Pr
  *
  * Throws INTERNAL_SERVER_ERROR if the JSON is unparseable or the verdict is not
  * one of the four allowed values. The evaluate procedure propagates this to the
- * client (FR-12).
+ * client.
  */
 function parseLlmResponse(rawContent: string): z.infer<typeof evaluateOutputSchema> {
 	// Strip possible markdown code fences the LLM may add despite instructions
@@ -307,8 +282,6 @@ function parseLlmResponse(rawContent: string): z.infer<typeof evaluateOutputSche
 	return result.data;
 }
 
-// ─── LLM Config Fetcher ───────────────────────────────────────────────────────
-
 /**
  * Fetches the LLM configuration from the database.
  * Throws NOT_FOUND if no configuration has been set up.
@@ -333,9 +306,6 @@ async function fetchLlmConfig(): Promise<LlmConfig> {
 
 	return config;
 }
-
-// ─── HTTP Error Map ───────────────────────────────────────────────────────────
-// Same error codes as ai-tutor.router.ts to stay consistent across the platform.
 
 const errorMap: Record<number, { code: TRPCError["code"]; message: string }> = {
 	400: { code: "BAD_REQUEST", message: "Invalid request to LLM server" },
